share getResources
